Moved TableViewSet to its own file.
