+++ mathesar/urls.py
@@ -4,10 +4,11 @@ from rest_framework_nested import routers
 from mathesar import views
 from mathesar.api import legacy_viewsets
 from mathesar.api.viewsets.schemas import SchemaViewSet
+from mathesar.api.viewsets.tables import TableViewSet
 
 
 router = routers.DefaultRouter()
-router.register(r'tables', legacy_viewsets.TableViewSet, basename='table')
+router.register(r'tables', TableViewSet, basename='table')
 router.register(r'schemas', SchemaViewSet, basename='schema')
 router.register(r'databases', legacy_viewsets.DatabaseViewSet, basename='database')
 router.register(r'data_files', legacy_viewsets.DataFileViewSet, basename='data-file')
